flytt søknadinnhold klasser inn i egne filer
