+++ database/src/main/kotlin/no/nav/su/se/bakover/database/søknad/SøknadRepoInternal.kt
@@ -11,9 +11,9 @@ import no.nav.su.se.bakover.database.hentListe
 import no.nav.su.se.bakover.database.tidspunkt
 import no.nav.su.se.bakover.domain.NavIdentBruker
 import no.nav.su.se.bakover.domain.Søknad
-import no.nav.su.se.bakover.domain.SøknadInnhold
 import no.nav.su.se.bakover.domain.journal.JournalpostId
 import no.nav.su.se.bakover.domain.oppgave.OppgaveId
+import no.nav.su.se.bakover.domain.søknadinnhold.SøknadInnhold
 import java.util.UUID
 
 internal object SøknadRepoInternal {
