+++ database/src/test/kotlin/no/nav/su/se/bakover/database/nøkkeltall/NøkkeltallPostgresRepoTest.kt
@@ -3,9 +3,9 @@ package no.nav.su.se.bakover.database.nøkkeltall
 import io.kotest.matchers.shouldBe
 import no.nav.su.se.bakover.database.TestDataHelper
 import no.nav.su.se.bakover.database.withMigratedDb
-import no.nav.su.se.bakover.domain.ForNav
 import no.nav.su.se.bakover.domain.SøknadInnholdTestdataBuilder
 import no.nav.su.se.bakover.domain.nøkkeltall.Nøkkeltall
+import no.nav.su.se.bakover.domain.søknadinnhold.ForNav
 import no.nav.su.se.bakover.test.fixedLocalDate
 import org.junit.jupiter.api.Test
