+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/SøknadInnholdTestdataBuilder.kt
@@ -5,11 +5,26 @@ import no.nav.su.se.bakover.common.februar
 import no.nav.su.se.bakover.common.januar
 import no.nav.su.se.bakover.common.juli
 import no.nav.su.se.bakover.domain.søknadinnhold.Boforhold
+import no.nav.su.se.bakover.domain.søknadinnhold.Ektefelle
 import no.nav.su.se.bakover.domain.søknadinnhold.EktefellePartnerSamboer
+import no.nav.su.se.bakover.domain.søknadinnhold.Flyktningsstatus
+import no.nav.su.se.bakover.domain.søknadinnhold.ForNav
 import no.nav.su.se.bakover.domain.søknadinnhold.Formue
+import no.nav.su.se.bakover.domain.søknadinnhold.HarSøktAlderspensjon
+import no.nav.su.se.bakover.domain.søknadinnhold.InnlagtPåInstitusjon
+import no.nav.su.se.bakover.domain.søknadinnhold.InntektOgPensjon
 import no.nav.su.se.bakover.domain.søknadinnhold.Kjøretøy
 import no.nav.su.se.bakover.domain.søknadinnhold.OppgittAdresse
 import no.nav.su.se.bakover.domain.søknadinnhold.Oppholdstillatelse
+import no.nav.su.se.bakover.domain.søknadinnhold.OppholdstillatelseAlder
+import no.nav.su.se.bakover.domain.søknadinnhold.PensjonsOrdningBeløp
+import no.nav.su.se.bakover.domain.søknadinnhold.Personopplysninger
+import no.nav.su.se.bakover.domain.søknadinnhold.SøknadsinnholdAlder
+import no.nav.su.se.bakover.domain.søknadinnhold.SøknadsinnholdUføre
+import no.nav.su.se.bakover.domain.søknadinnhold.TrygdeytelseIUtlandet
+import no.nav.su.se.bakover.domain.søknadinnhold.Uførevedtak
+import no.nav.su.se.bakover.domain.søknadinnhold.Utenlandsopphold
+import no.nav.su.se.bakover.domain.søknadinnhold.UtenlandsoppholdPeriode
 import java.time.LocalDate
 import java.time.Month.JANUARY
