+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/søknadinnhold/Ektefelle.kt
@@ -1,0 +1,3 @@
+package no.nav.su.se.bakover.domain.søknadinnhold
+
+data class Ektefelle(val formue: Formue, val inntektOgPensjon: InntektOgPensjon)
