+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/søknadinnhold/HarSøktAlderspensjon.kt
@@ -1,0 +1,5 @@
+package no.nav.su.se.bakover.domain.søknadinnhold
+
+data class HarSøktAlderspensjon(
+    val harSøktAlderspensjon: Boolean
+)
