+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/søknadinnhold/OppholdstillatelseAlder.kt
@@ -1,0 +1,6 @@
+package no.nav.su.se.bakover.domain.søknadinnhold
+
+data class OppholdstillatelseAlder(
+    val eøsborger: Boolean?,
+    val familiegjenforening: Boolean?
+)
