+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/søknadinnhold/Personopplysninger.kt
@@ -1,0 +1,7 @@
+package no.nav.su.se.bakover.domain.søknadinnhold
+
+import no.nav.su.se.bakover.domain.Fnr
+
+data class Personopplysninger(
+    val fnr: Fnr
+)
