+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/søknadinnhold/Utenlandsopphold.kt
@@ -1,0 +1,13 @@
+package no.nav.su.se.bakover.domain.søknadinnhold
+
+import java.time.LocalDate
+
+data class Utenlandsopphold(
+    val registrertePerioder: List<UtenlandsoppholdPeriode>? = null,
+    val planlagtePerioder: List<UtenlandsoppholdPeriode>? = null
+)
+
+data class UtenlandsoppholdPeriode(
+    val utreisedato: LocalDate,
+    val innreisedato: LocalDate
+)
