+++ domain/src/test/kotlin/no/nav/su/se/bakover/domain/søknad/SøknadTest.kt
@@ -1,8 +1,8 @@
 package no.nav.su.se.bakover.domain.søknad
 
 import io.kotest.matchers.shouldBe
-import no.nav.su.se.bakover.domain.ForNav
 import no.nav.su.se.bakover.domain.Søknad
+import no.nav.su.se.bakover.domain.søknadinnhold.ForNav
 import no.nav.su.se.bakover.test.fixedLocalDate
 import no.nav.su.se.bakover.test.fixedTidspunkt
 import no.nav.su.se.bakover.test.søknadinnhold
