+++ service/src/main/kotlin/no/nav/su/se/bakover/service/søknad/SøknadService.kt
@@ -5,8 +5,8 @@ import no.nav.su.se.bakover.common.persistence.SessionContext
 import no.nav.su.se.bakover.domain.NavIdentBruker
 import no.nav.su.se.bakover.domain.Saksnummer
 import no.nav.su.se.bakover.domain.Søknad
-import no.nav.su.se.bakover.domain.SøknadInnhold
 import no.nav.su.se.bakover.domain.journal.JournalpostId
+import no.nav.su.se.bakover.domain.søknadinnhold.SøknadInnhold
 import java.util.UUID
 
 interface SøknadService {
