+++ web/src/main/kotlin/no/nav/su/se/bakover/web/routes/søknad/søknadinnholdJson/BoforholdJson.kt
@@ -2,10 +2,10 @@ package no.nav.su.se.bakover.web.routes.søknad.søknadinnholdJson
 
 import arrow.core.Either
 import arrow.core.left
-import no.nav.su.se.bakover.domain.InnlagtPåInstitusjon
 import no.nav.su.se.bakover.domain.søknadinnhold.Boforhold
 import no.nav.su.se.bakover.domain.søknadinnhold.EktefellePartnerSamboer
 import no.nav.su.se.bakover.domain.søknadinnhold.FeilVedOpprettelseAvBoforhold
+import no.nav.su.se.bakover.domain.søknadinnhold.InnlagtPåInstitusjon
 import no.nav.su.se.bakover.domain.søknadinnhold.OppgittAdresse
 
 data class BoforholdJson(
