+++ web/src/main/kotlin/no/nav/su/se/bakover/web/routes/søknad/søknadinnholdJson/PersonopplysningerJson.kt
@@ -1,7 +1,7 @@
 package no.nav.su.se.bakover.web.routes.søknad.søknadinnholdJson
 
 import no.nav.su.se.bakover.domain.Fnr
-import no.nav.su.se.bakover.domain.Personopplysninger
+import no.nav.su.se.bakover.domain.søknadinnhold.Personopplysninger
 
 data class PersonopplysningerJson(val fnr: String) {
