+++ web/src/main/kotlin/no/nav/su/se/bakover/web/routes/søknad/søknadinnholdJson/SøknadsinnholdJson.kt
@@ -6,16 +6,16 @@ import arrow.core.left
 import arrow.core.right
 import com.fasterxml.jackson.annotation.JsonSubTypes
 import com.fasterxml.jackson.annotation.JsonTypeInfo
-import no.nav.su.se.bakover.domain.Flyktningsstatus
-import no.nav.su.se.bakover.domain.HarSøktAlderspensjon
-import no.nav.su.se.bakover.domain.OppholdstillatelseAlder
-import no.nav.su.se.bakover.domain.SøknadInnhold
-import no.nav.su.se.bakover.domain.SøknadsinnholdAlder
-import no.nav.su.se.bakover.domain.SøknadsinnholdUføre
-import no.nav.su.se.bakover.domain.Uførevedtak
 import no.nav.su.se.bakover.domain.søknadinnhold.FeilVedOpprettelseAvBoforhold
 import no.nav.su.se.bakover.domain.søknadinnhold.FeilVedOpprettelseAvFormue
 import no.nav.su.se.bakover.domain.søknadinnhold.FeilVedOpprettelseAvOppholdstillatelse
+import no.nav.su.se.bakover.domain.søknadinnhold.Flyktningsstatus
+import no.nav.su.se.bakover.domain.søknadinnhold.HarSøktAlderspensjon
+import no.nav.su.se.bakover.domain.søknadinnhold.OppholdstillatelseAlder
+import no.nav.su.se.bakover.domain.søknadinnhold.SøknadInnhold
+import no.nav.su.se.bakover.domain.søknadinnhold.SøknadsinnholdAlder
+import no.nav.su.se.bakover.domain.søknadinnhold.SøknadsinnholdUføre
+import no.nav.su.se.bakover.domain.søknadinnhold.Uførevedtak
 import no.nav.su.se.bakover.web.routes.søknad.søknadinnholdJson.BoforholdJson.Companion.toBoforholdJson
 import no.nav.su.se.bakover.web.routes.søknad.søknadinnholdJson.EktefelleJson.Companion.toJson
 import no.nav.su.se.bakover.web.routes.søknad.søknadinnholdJson.ForNavJson.Companion.toForNavJson
